Add ignorable module
